backend/requirements.txt
# FastAPI Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database & Auth
supabase>=2.3.4
python-jose[cryptography]>=3.3.0

# Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Blockchain
web3>=6.14.0
eth-account>=0.10.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Twilio
twilio>=8.10.0

# AI
groq>=0.4.2
openai>=1.9.0

# Voice
elevenlabs>=0.2.27

# Utils
python-dotenv>=1.0.0
phonenumbers>=8.13.27
aiofiles>=23.2.1

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
